Add files via upload
This project is my capstone project solution for the Microsoft Professional program in Data Science Course.

The data visualization was done in two parts for a thorough visualization and understanding of the data.

The Machine Learning models created ere XGBoost and Catboost. The catboost algorithm leverages its performance on only categorical features, either encoded or not. The XGboost algorithm works with both numerical and categorical columns, but the categorical columns need to be encoded. For this project, I target encoded all the categorical columns for a good performance of the XGboost model and I ensured overfitting (the problem of target encoding) as well accounted for.

The files could not be uploaded because of the large size even after compression.